fix: fixed bug
- call total number of row in table instead of the result's length
- fixed type to long

+++ src/main/java/me/h_yang/my_homepage/dto/BlogPostList.java
@@ -2,4 +2,4 @@ package me.h_yang.my_homepage.dto;
 
 import java.util.List;
 
-public record BlogPostList(List<BriefBlogPostDTO> postList, int totalCount) {}
+public record BlogPostList(List<BriefBlogPostDTO> postList, long totalCount) {}
